n less than zero, return

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -12,7 +12,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 
 	va_start(nums, n);
 
-	if (n < 0)
+	if ((int) n < 0)
 		return;
 
 	for (i = 0; i < (int) n - 1; i++)
